refactor(backend): split sync import runners
